refactor: split smartshare and nearby messages API
