+++ src/GoogleNearbyMessages.ts
@@ -1,58 +1,29 @@
 /*
- * Copyright (C) 2021 Newlogic Pte. Ltd.
+ * MIT License
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Copyright (c) 2020 Marc Rousavy
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 import { NativeModules, Platform, NativeEventEmitter } from 'react-native';
 import { useEffect, useState, useCallback, useReducer } from 'react';
-
-// ID PASS Smartshare
-const LINKING_ERROR =
-  `The package '@idpass/smartshare-react-native' doesn't seem to be linked. Make sure: \n\n` +
-  Platform.select({ ios: "- You have run 'pod install'\n", default: '' }) +
-  '- You rebuilt the app after installing the package\n' +
-  '- You are not using Expo managed workflow\n';
-
-const IdpassSmartshare = NativeModules.IdpassSmartshare
-  ? NativeModules.IdpassSmartshare
-  : new Proxy(
-      {},
-      {
-        get() {
-          throw new Error(LINKING_ERROR);
-        },
-      }
-    );
-
-IdpassSmartshare.noop; // to trigger iOS autolink
-
-interface HandlerFunc {
-  (message: string): void;
-}
-
-if (Platform.OS === 'android') {
-  const eventEmitter = new NativeEventEmitter();
-  IdpassSmartshare.handleNearbyEvents = (callback: HandlerFunc) =>
-    eventEmitter.addListener('EVENT_NEARBY', callback);
-  IdpassSmartshare.handleLogEvents = (callback: HandlerFunc) =>
-    eventEmitter.addListener('EVENT_LOG', callback);
-} else if (Platform.OS === 'ios') {
-  IdpassSmartshare.handleNearbyEvents = () => ({ remove: () => {} });
-  IdpassSmartshare.handleLogEvents = () => ({ remove: () => {} });
-}
-
-export default IdpassSmartshare;
 
 // Google Nearby Messages
 const { GoogleNearbyMessages } = NativeModules;
