+++ src/index.ts
@@ -1,0 +1,7 @@
+import * as GoogleNearbyMessages from './GoogleNearbyMessages';
+import IdpassSmartshare from './IdpassSmartshare';
+
+export default {
+  GoogleNearbyMessages,
+  IdpassSmartshare,
+};
